Vimrc refactor
Includes Fzf for finding files (and removing CtrlP), deoplete for
autocompletions, some typescript syntax files, and a new set of
functions for running tests, written from scratch because the old ones
that I lifted from someone else's vimrc, have stopped working across a
bunch of projects I use daily.

As part of this I've also removed the functions that deal with switching
from various parts of a rails application and replaced them with tpope's
Rails.vim, as yet I am not entirely sure on the benefits of this change
so I'll revisit it in the future.

Also some reshuffling of plugins (organising them into categories)